// src/interfaces/enum.ts
export enum LanguageType {
  FROM = 'from',
  TO = 'to'
}

export enum ActionType {
  INTERCHANGE_LANGUAGE = 'INTERCHANGE_LANGUAGE',
  SET_FROM_LANGUAGE = 'SET_FROM_LANGUAGE',
  SET_TO_LANGUAGE = 'SET_TO_LANGUAGE',
  SET_FROM_TEXT = 'SET_FROM_TEXT',
  SET_RESULT = 'SET_RESULT',
  SET_LOADING = 'SET_LOADING'
}